Moves app.py to root to keep consistency between tests and app execution
