+++ .circleci/config.yml
@@ -26,7 +26,7 @@ jobs:
       - run:
           name: "Run linter"
           command: |
-            pipenv run python -m flake8 stand_up_bot/ tests/stand_up_bot/
+            pipenv run python -m flake8 app.py stand_up_bot/ tests/stand_up_bot/
       - run:
           name: "Run tests"
           command: |
