+++ Procfile
@@ -1,1 +1,1 @@
-web: python stand_up_bot/app.py
+web: python app.py
